% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carry_absorbing_forward.R
\name{carry_absorbing_forward}
\alias{carry_absorbing_forward}
\title{Carry absorbing state forward for ordinal longitudinal data}
\usage{
carry_absorbing_forward(df, absorb, tmax)
}
\arguments{
\item{df}{Input data frame}

\item{absorb}{Which state is absorbing? Integer.}

\item{tmax}{What is the last day of the trial? Integer.}
}
\value{
Data frame with absorbing state carried forward
}
\description{
Carry absorbing state forward for ordinal longitudinal data
}
