% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sop.R
\name{compute_sop}
\alias{compute_sop}
\title{Generate the SOP matrix implied by the OTM parameters}
\usage{
compute_sop(params, covariates, baseline_y, times, absorb, output_type = "df")
}
\arguments{
\item{params}{A named list containing the parameters of the OTM.
Must include "beta_yprev": k-1 parameters for the k previous states,
"beta_regression": parameters for each covariates,
"beta_time": a single slope for the time parameter, and
"cutpoints": the cutpoint parameters.}

\item{covariates}{A vector of covariates}

\item{baseline_y}{The baseline state of the subject}

\item{times}{A vector of times to generate data for}

\item{absorb}{Is the last y state an absorbing state?}

\item{output_type}{Either "df" or "matrix"}
}
\value{
SOPs as a data frame or matrix.
}
\description{
Generate the SOP matrix implied by the OTM parameters
}
