% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_transition_matrix.R
\name{compute_transition_matrix}
\alias{compute_transition_matrix}
\title{Compute the transition matrix implied by the OTM parameters on a given day}
\usage{
compute_transition_matrix(params, day, covariates, absorb)
}
\arguments{
\item{params}{A named list containing the parameters of the OTM.
Must include "beta_yprev": k-1 parameters for the k previous states,
"beta_regression": parameters for each covariates,
"beta_time": a single slope for the time parameter, and
"cutpoints": the cutpoint parameters.}

\item{day}{Study day}

\item{covariates}{A vector of covariates}

\item{absorb}{Is the last y state an absorbing state?}
}
\value{
Matrix of transition probabilities
}
\description{
Compute the transition matrix implied by the OTM parameters on a given day
}
