% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_transition_probabilities.R
\name{compute_transition_probabilities}
\alias{compute_transition_probabilities}
\title{Compute a vector of transition probability from an OTM}
\usage{
compute_transition_probabilities(params, yprev, day, covariates)
}
\arguments{
\item{params}{A named list containing the parameters of the OTM. Must include "beta_yprev": k-1 parameters for the k previous states,
"beta_regression": parameters for each covariates, "beta_time": a single slope for the time parameter, and
"cutpoints": the cutpoint parameters.}

\item{yprev}{The previous state to be conditioned on.}

\item{day}{The day to compute the transition probabilities for.}

\item{covariates}{A vector of covariates}
}
\value{
A vector of transition probabilities. The length is equal to the number of states.
}
\description{
Given the parameters from an ordinal transition model (OTM), calculate the transition probabilities for a given time and previous state.
}
