% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_improvement.R
\name{create_improvement}
\alias{create_improvement}
\title{Generate the time to improvement dataset from an OTM dataset}
\usage{
create_improvement(df, death_state, threshold, tmax)
}
\arguments{
\item{df}{An OTM dataset generated with otm::generate_dataset}

\item{death_state}{Which state corresponds to death? Must be integer.}

\item{threshold}{What is the threshold for improvement?}

\item{tmax}{The final day of the trial (used for censoring at death)}
}
\value{
A survival dataset
}
\description{
Generate the time to improvement dataset from an OTM dataset
}
