% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_trajectory.R
\name{generate_dataset}
\alias{generate_dataset}
\title{TBD}
\usage{
generate_dataset(
  cutpoints,
  beta_yprev,
  beta_t,
  beta_tx,
  beta_t_tx,
  tx_end,
  baseline_y,
  times,
  tx_type,
  absorb,
  n_subjects
)
}
\description{
TBD
}
