% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sop_plot.R
\name{generate_sop_plot}
\alias{generate_sop_plot}
\title{Plot the SOPs implied by the OTM parameters}
\usage{
generate_sop_plot(sop_dataframe, type = "both")
}
\arguments{
\item{sop_dataframe}{A SOP data.frame generated by otm::compute_sop()}

\item{type}{Use "bar", "line", or "both" for the plot}
}
\value{
SOP plot using ggplot2
}
\description{
Plot the SOPs implied by the OTM parameters
}
