% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm-package.R
\docType{package}
\name{otm-package}
\alias{otm}
\alias{otm-package}
\title{otm: Functions for Ordinal Transition Models}
\description{
TBD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maxdrohde.github.io/otm/}
}

}
\author{
\strong{Maintainer}: Maximilian Rohde \email{maxdrohde@gmail.com} (\href{https://orcid.org/0000-0002-3731-3372}{ORCID})

}
\keyword{internal}
