% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm_verify.R
\name{otm_verify}
\alias{otm_verify}
\title{Verify ordinal longitudinal dataset}
\usage{
otm_verify(df)
}
\arguments{
\item{df}{data.frame to check}
}
\value{
Original data frame or error message
}
\description{
This function does three things:
\itemize{
\item Check that \code{id}, \code{t}, \code{yprev}, \code{t}, and \code{tx} columns are in the dataset
\item Standardize the data types for the columns
\item Sort the data by id and t
}
}
