% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOP_from_matrix_list.R
\name{SOP_from_matrix_list}
\alias{SOP_from_matrix_list}
\title{Compute an SOP from a list of transition matrices}
\usage{
SOP_from_matrix_list(list_of_matrices, baseline_state, output_type = "df")
}
\arguments{
\item{list_of_matrices}{A list of transition matrices, one for each time.}

\item{baseline_state}{The baseline state that the SOP is conditioned on.}

\item{output_type}{What format should be used for the output? Either "matrix" or "df".}
}
\value{
Either a matrix or data.frame containing the SOPs
}
\description{
Let the transition matrices be denoted P1, P2, ... Pn for times 1 to n.

The SOP at time j, starting from a baseline state of k, is P_t * P_(t-1) * ... * P_2 * SOP_(t = 1).

SOP_(t = 1) is the kth column of P_1.
}
\examples{
\dontrun{
SOP_from_matrix_list(l, baseline_state = 4, output_type = "matrix")
}
}
