% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_transition_probabilities}
\alias{compute_transition_probabilities}
\title{Compute a vector of transition probability from an OTM}
\usage{
compute_transition_probabilities(params, yprev, day, treatment)
}
\arguments{
\item{params}{A named list containing the parameters of the OTM.
Must include "beta_yprev": k-1 parameters for the k previous states,
"beta_treatment": parameter for the treatment effect,
"beta_time": a single slope for the time parameter, and
"cutpoints": the cutpoint parameters.}

\item{yprev}{The previous state to be conditioned on.}

\item{day}{The day to compute the transition probabilities for.}

\item{treatment}{Treatment group? 0 or 1.}
}
\value{
A vector of transition probabilities.
}
\description{
Given the parameters from an ordinal transition model (OTM),
calculate the transition probabilities for a given time and previous state.
}
