% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survival.R
\name{create_survival}
\alias{create_survival}
\title{Generate the survival dataset from an OTM dataset}
\usage{
create_survival(df, death_state, recovery_states, tmax)
}
\arguments{
\item{df}{An OTM dataset generated with otm::generate_dataset}

\item{death_state}{Which state corresponds to death? Must be integer.}

\item{recovery_states}{Which state(s) correspond to recovery? Must be integer.}

\item{tmax}{The final day of the trial (used for censoring at death)}
}
\value{
A survival dataset
}
\description{
Generate the survival dataset from an OTM dataset
}
