% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_record.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate a dataset of OTM data}
\usage{
generate_dataset(params, covariates, baseline_y, times, absorb, n_subjects)
}
\arguments{
\item{params}{A named list containing the parameters of the OTM. Must include "beta_yprev": k-1 parameters for the k previous states,
"beta_regression": parameters for each covariates, "beta_time": a single slope for the time parameter, and
"cutpoints": the cutpoint parameters.}

\item{covariates}{A vector of covariates}

\item{baseline_y}{The baseline state of the subjects used when generating the data}

\item{times}{A vector of times to generate data for}

\item{absorb}{Which state is absorbing?}

\item{n_subjects}{How many subject records should be generated?}
}
\value{
A data frame of data generated under the first order OTM model.
}
\description{
Generate a dataset of OTM data
}
