% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_dataset}
\alias{generate_dataset}
\title{Generate a dataset of OTM data (with C++)}
\usage{
generate_dataset(params, treatment, baseline_y, times, absorb, n_subjects)
}
\arguments{
\item{params}{A named list containing the parameters of the OTM.
Must include "beta_yprev": k-1 parameters for the k previous states,
"beta_treatment": parameter for the treatment effect,
"beta_time": a single slope for the time parameter, and
"cutpoints": the cutpoint parameters.}

\item{treatment}{Treatment group? 0 or 1.}

\item{baseline_y}{The baseline state of the subjects used when generating the data}

\item{times}{A vector of times to generate data for}

\item{absorb}{Is the last state absorbing?}

\item{n_subjects}{How many subject records should be generated?}
}
\value{
A data frame of data generated under the first order OTM model.
}
\description{
Generate a dataset of OTM data (with C++)
}
