% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sop_plot.R
\name{generate_sop_plot}
\alias{generate_sop_plot}
\title{Plot the SOPs implied by the OTM parameters}
\usage{
generate_sop_plot(sop_dataframe)
}
\arguments{
\item{sop_dataframe}{A}
}
\value{
Matrix of transition probabilities
}
\description{
Plot the SOPs implied by the OTM parameters
}
